Add a custom restartsrv module.
There are a bunch of reasons for this module, both a lack of
restart feature on CentOS 6, to the inability of restartsrv
to call systemctl reload, as systemctl restart calls stop and
then start.

The behavior on both C6 and C7 is now consistent and performs
as you would expect.