Update tests to be compatible with latest Test::MockFile version
Case EA-10519: Test::MockFile was updated to the latest version via
rpm.versions.  This resulted in several unit tests failing due to
changes to Test::MockFile.  Namely, strict mode was enabled by default
for Test::MockFile and our tests were not written with this in mind or
to be "small" tests.  There were also changes to the 'dir' class method in
Test::MockFile that required some minor tweaks to our tests.

Changelog: