EA-10751: Leave X-Forwarded-Host blank for service subdomains
The X-Forwarded-Host header will be set again by Apache when
forwarding to cpsrvd.  If nginx has set this already, Apache
will simply add it again to X-Forwarded-Host causing the header
seen by cpsrvd to be set as something like:
"cpanel.cptest.tld, cpanel.cptest.tld".  This will cause
Cpanel::Server::HTTP to reject it has having bad syntax
and substitute it for the main server hostname; this, in turn,
causes the error described in the case.