EA-10999: Update logic to check if service is listening on ports 80 or 443
In 110, Cpanel::AppPort::get_pids_bound_to_ports() was refactored
and, as a result, the "process" name returned by it is no
longer simply the name of the executable as it had been before.
It now comes from /proc/<pid>/cmdline.  It may contain the full
path to the exe, or something like "nginx: logger process".  So,
007-restartsrv_nginx needs to switch from doing direct comparison
to the "process" value in the hash to using a regular expression.